// src/models/TMDBModel.ts
export interface MovieQuery {
  adult: boolean;
  backdrop_path: string;
  genre_ids: number[];
  id: number;
  original_language: string;
  original_title: string;
  overview: string;
  popularity: number;
  poster_path: string;
  release_date: string;
  title: string;
  video: boolean;
  vote_average: number;
  vote_count: number;
}

export interface APIQuery {
  page: number;
  results: MovieQuery[];
  total_pages: number;
  total_results: number;
}

export interface TMDBMovie {
  adult: boolean;
  backdrop_path: string | null;
  belongs_to_collection: null | {
    id: number;
    name: string;
    poster_path: string;
    backdrop_path: string;
  };
  budget: number;
  genres: {
    id: number;
    name: string;
  }[];
  homepage: string | null;
  id: number;
  imdb_id: string | null;
  original_language: string;
  original_title: string;
  overview: string | null;
  popularity: number;
  poster_path: string | null;
  production_companies: {
    id: number;
    logo_path: string | null;
    name: string;
    origin_country: string;
  }[];
  production_countries: {
    iso_3166_1: string;
    name: string;
  }[];
  release_date: string;
  revenue: number;
  runtime: number | null;
  spoken_languages: {
    english_name: string;
    iso_639_1: string;
    name: string;
  }[];
  status:
    | string
    | "Rumored"
    | "Planned"
    | "In Production"
    | "Post Production"
    | "Released"
    | "Canceled";

  tagline: string | null;
  title: string;
  video: boolean;
  vote_average: number;
  vote_count: number;
}

export interface TMDBWatchProviders {
  link: string;
  rent: {
    display_priority: number;
    logo_path: string;
    provider_id: number;
    provider_name: string;
  }[];
  buy: {
    display_priority: number;
    logo_path: string;
    provider_id: number;
    provider_name: string;
  }[];
  flatrate: {
    display_priority: number;
    logo_path: string;
    provider_id: number;
    provider_name: string;
  }[];
}